refactor: 로고 아이콘 위치를 일괸되게 수정
로고 아이콘을 다른 로고와 동일하게 svgs/logo 하위로 이동시킴
