--- src/shared/assets/svgs/index.ts
@@ -1,1 +1,0 @@
-export { default as LogoIcon } from './LogoIcon';
