+++ src/shared/assets/svgs/logo/LogoIcon.tsx
@@ -3,12 +3,12 @@ import type { SVGProps } from 'react';
 const SvgLogoIcon = (props: SVGProps<SVGSVGElement>) => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
-    fill="current"
+    fill="none"
     viewBox="0 0 98 36"
     {...props}
   >
     <path
-      fill="current"
+      fill="#292C30"
       d="M2.856 28c-1.368 0-2.064-.768-1.872-2.136L2.88 12.496c.192-1.224.912-1.872 2.16-1.872h2.256c1.224 0 1.944.6 2.256 1.776l2.784 10.536L15.12 12.4c.312-1.176 1.032-1.776 2.232-1.776h2.256c1.248 0 1.992.648 2.16 1.872l1.896 13.368C23.856 27.232 23.16 28 21.792 28h-.816c-1.248 0-1.968-.648-2.136-1.896l-1.2-9.12-2.568 9.288C14.76 27.424 14.016 28 12.816 28h-.984c-1.2 0-1.92-.576-2.256-1.728l-2.568-9.288-1.2 9.12C5.64 27.352 4.92 28 3.672 28zm30.808.192c-4.392 0-8.256-2.544-8.256-8.856 0-6.192 3.912-8.832 8.304-8.832 4.32 0 8.304 2.64 8.304 8.832 0 6.312-3.96 8.856-8.352 8.856m0-4.224c2.064 0 3.48-1.56 3.48-4.632 0-2.952-1.416-4.584-3.456-4.584-2.112 0-3.504 1.632-3.504 4.584 0 3.072 1.368 4.632 3.48 4.632M46.822 28c-1.32 0-2.04-.696-2.04-2.016v-13.32c0-1.32.72-2.04 2.04-2.04h3.336c7.008 0 9.624 3.168 9.624 8.712 0 5.448-2.616 8.664-9.624 8.664zm3.528-4.272c3.648 0 4.536-1.536 4.536-4.44 0-2.856-.888-4.416-4.56-4.416h-.672v8.856zM64.611 28c-1.32 0-2.04-.696-2.04-2.016v-13.32c0-1.32.72-2.04 2.04-2.04h3.336c7.008 0 9.624 3.168 9.624 8.712 0 5.448-2.616 8.664-9.624 8.664zm3.528-4.272c3.648 0 4.536-1.536 4.536-4.44 0-2.856-.888-4.416-4.56-4.416h-.672v8.856zm19.83 4.464c-4.393 0-8.257-2.544-8.257-8.856 0-6.192 3.912-8.832 8.304-8.832 4.32 0 8.304 2.64 8.304 8.832 0 6.312-3.96 8.856-8.352 8.856m0-4.224c2.063 0 3.48-1.56 3.48-4.632 0-2.952-1.417-4.584-3.457-4.584-2.112 0-3.504 1.632-3.504 4.584 0 3.072 1.368 4.632 3.48 4.632"
     />
   </svg>
