+++ src/shared/assets/svgs/logo/index.ts
@@ -1,2 +1,3 @@
 export { default as Slack } from './Slack';
 export { default as Kakaotalk } from './Kakaotalk';
+export { default as LogoIcon } from './LogoIcon';
